Create letsupgrade aggignment2
letsupgrade c++ assignment2
Write a C++ program that contains a function called reverseString which takes a string as input and reverses it. The program should also include a main function that prompts the user to enter a string, calls the reverseString function to reverse the string, and then displays the reversed string.